Refactor: Simplify `openStudyOptionsActivity` method
This commit refactors the `openStudyOptionsActivity` method in `DeckPicker.kt`.
The `withDeckOptions` parameter, which was previously used to determine if the deck options activity should be loaded, has been removed. This simplifies the method and its call sites, as the deck options are now consistently handled by the `StudyOptionsComposeActivity`.

Additionally, the width of the `ModalDrawerSheet` in the navigation drawer has been explicitly set to `310.dp`.