Feat: Implement "More options" dropdown menu in Reviewer
This commit adds a functional "More options" dropdown menu to the floating toolbar in the Reviewer screen.

Previously, the "More options" icon was a placeholder. Now, clicking the icon reveals a `DropdownMenu` containing a list of actions a user can take during review, such as "Redo," "Edit note," and "Bury card." Each menu item is displayed with a corresponding icon and text.

The implementation uses a `DropdownMenu` composable triggered by an `IconButton`. The list of menu options is defined with their associated icons and text, though the `onClick` actions are yet to be implemented.