Feat: Wire up actions in Reviewer overflow menu
This commit connects UI events to view model actions for several items in the reviewer's bottom sheet menu.

Specifically, the following actions are now implemented:
- "Edit card"
- "Bury card"
- "Suspend card"
- "Mark note"

The data structure for menu options was also updated from a `List<Pair>` to a `List<Triple>` to include the action lambda for each item.