feat(CardBrowser): Ellipsize long sort-by column names
This commit improves the UI of the sort button in the Card Browser.

When the name of the column used for sorting is too long to fit in the available space, it will now be ellipsized (...) instead of wrapping to a new line or being cut off abruptly. This is achieved by setting `maxLines = 1` and `overflow = TextOverflow.Ellipsis` on the `Text` composable.