feat: Show selected deck name on empty browser screen
This commit enhances the "no cards found" message in the Card Browser to specify the name of the currently selected deck.

When a search returns no results, the screen now displays "No cards in deck: [Deck Name]" instead of a generic message. If all decks are selected, it will show "No cards in deck: All Decks". This provides better context to the user about their current search scope.